Reorganized routing code, fixed logout button, removed depreciated /login get route
