+++ server.js
@@ -11,51 +11,6 @@ var db = require("./models");
 
 var PORT = process.env.PORT || 3000;
 
-// Configure the local strategy for use by Passport.
-//
-// The local strategy require a `verify` function which receives the credentials
-// (`username` and `password`) submitted by the user.  The function must verify
-// that the password is correct and then invoke `cb` with a user object, which
-// will be set at `req.user` in route handlers after authentication.
-passport.use(
-  new Strategy(function(username, password, cb) {
-    findByUsername(username, function(err, user) {
-      if (err) {
-        return cb(err);
-      }
-      if (!user) {
-        return cb(null, false);
-      }
-      if (user.password !== password) {
-        console.log(`LOGIN for ${user.username} FAILED - INCORRECT PASSWORD`);
-        return cb(null, false);
-      }
-      console.log("LOGIN SUCCESSFUL!!!");
-      return cb(null, user);
-    });
-  })
-);
-
-// Configure Passport authenticated session persistence.
-//
-// In order to restore authentication state across HTTP requests, Passport needs
-// to serialize users into and deserialize users out of the session.  The
-// typical implementation of this is as simple as supplying the user ID when
-// serializing, and querying the user record by ID from the database when
-// deserializing.
-passport.serializeUser(function(user, cb) {
-  cb(null, user.id);
-});
-
-passport.deserializeUser(function(id, cb) {
-  findById(id, function(err, user) {
-    if (err) {
-      return cb(err);
-    }
-    cb(null, user);
-  });
-});
-
 var app = express();
 // Handlebars
 app.engine(
@@ -87,55 +42,6 @@ app.use(passport.session());
 //app.use(express.json());
 app.use(express.static("public"));
 
-app.get("/", function(req, res) {
-  res.render("index", { user: req.user });
-});
-
-app.get("/login", function(req, res) {
-  console.log("Rendering login page");
-  res.render("login");
-});
-
-app.get("/logout", function(req, res) {
-  console.log("LOG OUT INVOKED");
-  req.logout();
-  res.redirect("/");
-});
-
-app.get("/profile", require("connect-ensure-login").ensureLoggedIn(), function(
-  req,
-  res
-) {
-  console.log(req.user);
-  db.bike
-    .findAll({
-      where: {
-        username: req.user.username
-      }
-    })
-    .then(bikeData => {
-      let bikeList = [];
-      bikeData.forEach(bike => {
-        // cleans up the results to parse easier
-        bikeList.push(bike.dataValues);
-      });
-
-      const profileData = { user: req.user, bikes: bikeList };
-      console.log("RENDERING PROFILE DATA:", profileData);
-      res.render("profile", profileData);
-    });
-});
-
-app.post(
-  "/login",
-  passport.authenticate("local", { failureRedirect: "/login" }),
-  function(req, res) {
-    console.log("REDIRECTING NOW");
-    res.redirect("/profile");
-    //res.render("/profile");
-  }
-);
-
 // Routes
 //require("./routes/apiRoutes")(app);
 require("./routes/htmlRoutes")(app);
@@ -144,8 +50,6 @@ var syncOptions = {
   force: true
 };
 
-// If running a test, set syncOptions.force to true
-// clearing the `testdb`
 if (process.env.NODE_ENV === "test") {
   syncOptions.force = true; // I've set force sync true here to make development easier - in case we update Schema. -Scott
 }
@@ -238,42 +142,3 @@ db.sequelize.sync(syncOptions).then(() => {
 });
 
 module.exports = app;
-
-function findByUsername(username, cb) {
-  db.user
-    .findAll({
-      where: {
-        username: username
-      }
-    })
-    .then(data => {
-      const record = data[0].dataValues;
-      console.log("________________________");
-      console.log(record);
-      console.log("________________________");
-      if (username === record.username) {
-        console.log("FOUND USER:", username);
-        return cb(null, record);
-      } else {
-        console.log("USER NOT FOUND.");
-        return cb(null, null);
-      }
-    });
-}
-
-function findById(id, cb) {
-  db.user
-    .findAll({
-      where: {
-        id: id
-      }
-    })
-    .then(data => {
-      const user = data[0].dataValues;
-      if (user) {
-        cb(null, user);
-      } else {
-        cb(new Error("User does not exist"));
-      }
-    });
-}
