+++ views/index.handlebars
@@ -7,7 +7,7 @@
 			<h1>{{msg}}</h1>
 			<h1 class="animated fadeInDown delay-07s">Welcome to The Bike Shepherd</h1>
 			<ul class="we-create animated fadeInUp delay-1s">
-				<li>Register your bike to prevent it from being stolen. Empower yourself today.</li>
+				<li>Register your bike to help fight theft. Empower yourself today.</li>
 			</ul>
 
 				<!--modal: login / register form-->
@@ -40,17 +40,16 @@
 									<form action="/login" method="post">
 										<div class="modal-body mb-1">
 											<div class="md-form form-sm mb-5">
-												<i class="fas fa-envelope prefix"></i>
-												<input type="text" name="username"/>
+												<label data-error="wrong" data-success="right" for="modalLRInput10">Your Username:</label>
+												<input type="text" name="username" class="form-control form-control-sm validate"/>
 												{{!-- <input type="text" id="username" class="form-control form-control-sm validate"> --}}
-												<label data-error="wrong" data-success="right" for="modalLRInput10">Your Username</label>
+												
 											</div>
-
 											<div class="md-form form-sm mb-4">
-												<i class="fas fa-lock prefix"></i>
-												<input type="password" name="password"/>
+												<label data-error="wrong" data-success="right" for="modalLRInput11">Your Password:</label>
+												<input type="password" name="password" class="form-control form-control-sm validate"/>
 												{{!-- <input type="password" id="password" class="form-control form-control-sm validate"> --}}
-												<label data-error="wrong" data-success="right" for="modalLRInput11">Your Password</label>
+												<i class="fas fa-lock prefix"></i>
 											</div>
 
 											<div class="text-center mt-2">
@@ -75,30 +74,37 @@
 								<div class="tab-pane fade" id="panel8" role="tabpanel">
 
 									<!--body-->
-									<div class="modal-body">
-										<div class="md-form form-sm mb-5">
-											<i class="fas fa-envelope prefix"></i>
-											<input type="email" id="modalLRInput12" class="form-control form-control-sm validate">
-											<label data-error="wrong" data-success="right" for="modalLRInput12">Your Email</label>
-										</div>
+									<form action="/register" method="post">
+										<div class="modal-body">
+											<div class="md-form form-sm mb-5">
+												<input type="text" name="username"/>
+												<label data-error="wrong" data-success="right" for="modalLRInput10">Your Username</label>
+											</div>
 
-										<div class="md-form form-sm mb-5">
-											<i class="fas fa-lock prefix"></i>
-											<input type="password" id="modalLRInput13" class="form-control form-control-sm validate">
-											<label data-error="wrong" data-success="right" for="modalLRInput13">Your Password</label>
-										</div>
+											<div class="md-form form-sm mb-5">
+												<i class="fas fa-envelope prefix"></i>
+												<input type="email" id="modalLRInput12" class="form-control form-control-sm validate">
+												<label data-error="wrong" data-success="right" for="modalLRInput12">Your Email</label>
+											</div>
 
-										<div class="md-form form-sm mb-4">
-											<i class="fas fa-lock prefix"></i>
-											<input type="password" id="modalLRInput14" class="form-control form-control-sm validate">
-											<label data-error="wrong" data-success="right" for="modalLRInput14">Repeat Password</label>
-										</div>
+											<div class="md-form form-sm mb-5">
+												<i class="fas fa-lock prefix"></i>
+												<input type="password" id="modalLRInput13" class="form-control form-control-sm validate">
+												<label data-error="wrong" data-success="right" for="modalLRInput13">Your Password</label>
+											</div>
 
-										<div class="text-center form-sm mt-2">
-											<button class="btn btn-info">Sign Up <i class="fas fa-sign-in ml-1"></i></button>
-										</div>
+											<div class="md-form form-sm mb-4">
+												<i class="fas fa-lock prefix"></i>
+												<input type="password" id="modalLRInput14" class="form-control form-control-sm validate">
+												<label data-error="wrong" data-success="right" for="modalLRInput14">Repeat Password</label>
+											</div>
 
-									</div>
+											<div class="text-center form-sm mt-2">
+												<button class="btn btn-info">Sign Up <i class="fas fa-sign-in ml-1"></i></button>
+											</div>
+
+										</div>
+									</form>
 									<!--footer-->
 									<div class="modal-footer">
 										<div class="options text-right">
