+++ views/partials/navbar_partial.handlebars
@@ -9,7 +9,7 @@
 
 			<li><a href="/team">Team</a></li>
 				<li><a href="/profile">Profile</a></li>
-			<li><a href="/" class="link animated fadeInUp delay-1s servicelink">Log Out</a></li>
+			<li><a href="/logout" class="link animated fadeInUp delay-1s servicelink">Log Out</a></li>
 		</ul>
 
 		<a class="res-nav_click" href="#"><i class="fas fa-bars"></i></a>
